Sprememba

+++ UrbaneMreze/Models/SpotsModel.cs
@@ -18,9 +18,9 @@ namespace UrbaneMreze.Models
 
         public String Description { get; set; }
 
-        public Decimal Longitude { get; set; }
+        public Double Longitude { get; set; }
 
-        public Decimal Latitude { get; set; }
+        public Double Latitude { get; set; }
 
         public DateTime DateCreated { get; set; }
